+++ blockchain/blockchain.conf
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+export _I32COIN_NUM_NEIGHBORS="4"
+export _I32COIN_HASH_SIZE="32"
+export _I32COIN_DIFFICULTY="31"
+export _I32COIN_REWARD="25"
