+++ go.mod
@@ -1,0 +1,11 @@
+module github.com/JMWorden/int32coin
+
+go 1.14
+
+require (
+	github.com/cbergoon/merkletree v0.2.0
+	github.com/ethereum/go-ethereum v1.9.14
+	github.com/libp2p/go-libp2p v0.9.2
+	github.com/libp2p/go-libp2p-core v0.5.6
+	golang.org/x/crypto v0.0.0-20200510223506-06a226fb4e37
+)
